use a minimal header and no footer for all pages but index
