+++ lib/components/layout/_layout.scss
@@ -2,6 +2,7 @@
 @import "./styles/header";
 @import "./styles/nav";
 @import "./styles/footer";
+@import "./styles/related";
 
 a[href="#main"] { // skip link
 	position: absolute;
