+++ lib/components/layout/macros/header.jsx
@@ -1,10 +1,22 @@
 import { assetURI } from "../../../../views/util";
 import { createElement } from "complate-stream";
 
-let logo = assetURI("faucet-logotype.png");
-let altLogo = assetURI("faucet-logotype.gif");
+function TinyHeader({ shortName, relatedLinks }) {
+	let logo = assetURI("faucet-logotype-monochrome.svg");
 
-export default function Header({ shortName, tagline }) {
+	return <header class="tiny-site-header">
+		<a href="/">
+			<img src={logo} alt={shortName} class="logo" />
+		</a>
+
+		{relatedLinks}
+	</header>;
+}
+
+function BigHeader({ shortName, tagline }) {
+	let logo = assetURI("faucet-logotype.png");
+	let altLogo = assetURI("faucet-logotype.gif");
+
 	return <header class="site-header">
 		<p class="tagline">{tagline}</p>
 
@@ -12,4 +24,9 @@ export default function Header({ shortName, tagline }) {
 			<img src={logo} alt={shortName} data-src={altLogo} />
 		</div>
 	</header>;
+}
+
+export default function Header({ shortName, tagline, relatedLinks, tiny }) {
+	return tiny ? <TinyHeader shortName={shortName} relatedLinks={relatedLinks}/> :
+		<BigHeader shortName={shortName} tagline={tagline}/>;
 }
