+++ lib/components/layout/styles/_header.scss
@@ -68,3 +68,17 @@
 		}
 	}
 }
+
+.tiny-site-header {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	max-height: 5rem;
+	margin-bottom: 2rem;
+	padding: 0.5rem 1.5rem;
+	background-color: $faucet-blue;
+
+	> a img {
+		width: 8rem;
+	}
+}
